refactor(analysis-types): extract effort cocomo DTO
