+++ crates/tokmd-analysis-types/src/effort.rs
@@ -6,6 +6,7 @@
 
 use serde::{Deserialize, Serialize};
 mod assumptions;
+mod cocomo;
 mod confidence;
 mod delta;
 mod driver;
@@ -14,6 +15,7 @@ mod results;
 mod size;
 
 pub use assumptions::EffortAssumptions;
+pub use cocomo::CocomoReport;
 pub use confidence::{EffortConfidence, EffortConfidenceLevel};
 pub use delta::{EffortDeltaClassification, EffortDeltaReport};
 pub use driver::{EffortDriver, EffortDriverDirection};
@@ -31,17 +33,4 @@ pub struct EffortEstimateReport {
     pub assumptions: EffortAssumptions,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub delta: Option<EffortDeltaReport>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CocomoReport {
-    pub mode: String,
-    pub kloc: f64,
-    pub effort_pm: f64,
-    pub duration_months: f64,
-    pub staff: f64,
-    pub a: f64,
-    pub b: f64,
-    pub c: f64,
-    pub d: f64,
 }
